Test clean. Rearrangement
